python-socket-io-server/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
Click==7.0
cycler==0.10.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
monotonic==1.5
multidict==4.7.5
numpy==1.18.1
pandas==1.0.1
pylsl==1.12.2
pyparsing==2.4.6
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
Werkzeug==1.0.0
yarl==1.4.2
scipy
